RSSFE-0 feat: add custom hook for mutation
